Update hf-diffusers.libsonnet to require Pillow >= 9.4.0 (#316)
This is to fix failing Pytorch/XLA diffusers tests. One of the dependencies requires the version of pillow to be greater than 9.4.0 https://github.com/huggingface/datasets/pull/6883/files
but we will have to wait for a new release of the dependency for a new version to be released. Temporarily fixing this by adding this line

Co-authored-by: Pei Zhang <piz@google.com>