Add a new DAG to test if the multi-host node pool availability is correctly reported during a node pool rollback
This change adds a new DAG which automates the process
of testing if the availability of a multi-host node pool
is correctly monitored while a node pool rollback takes
place.

This DAG requires an existing cluster. It will create a
node-pool under the specified cluster, and clean it up
after the tests.

---------

Co-authored-by: Alfred Yu <alfred.yu@cienet.com>